test(runtime): move bidirectional E2E test to integration binary; drop #[ignore]
Move `bidirectional_end_to_end_echo` from `push_router.rs::tests` (`#[ignore]`d due to the `GLOBAL_TCP_SERVER` cross-runtime bug) into `lib/runtime/tests/bidirectional_e2e.rs`.

Each file in `tests/` becomes its own cargo-test binary, which gets its own OS process. The `GLOBAL_TCP_SERVER` `OnceCell` is per-process, so the accept loop is bound to a runtime that lives for the entire test binary — no orphan-Arc-from-an-earlier-test situation. The test now runs unconditionally with `cargo test`, no `--include-ignored`, no manual gating.

Same test body, no behavior change. Uses the public crate surface (`dynamo_runtime::pipeline::*`, `engine::*`, `error::DynamoError`, etc.) instead of in-crate paths. The pieces this exercises — `Ingress::for_engine`, `PushRouter`, `RouterMode`, `Controller`, `ResponseStream` — are all already exposed by the runtime crate's public API. The original in-crate copy + its lengthy `#[ignore]` rationale are deleted from `push_router.rs::tests` (the rationale lives in the integration test's module doc comment and in the followup-global-tcp-server-runtime-lifetime note).
